Refactor and renaming and compilation
