+++ src/main/java/mktpromomarc/flotilla/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package mktpromomarc.flotilla;
 
 import org.apache.catalina.Context;
 import org.apache.catalina.LifecycleException;
