+++ src/main/java/mktpromomarc/flotilla/controller/LoginController.java
@@ -1,11 +1,11 @@
-package org.example.controller;
+package mktpromomarc.flotilla.controller;
 
 import com.google.gson.Gson;
-import org.example.dto.Credentials;
-import org.example.repository.UsuariosRepository;
-import org.example.security.Encoder;
-import org.example.security.Validator;
-import org.example.service.UsuariosService;
+import mktpromomarc.flotilla.dto.Credentials;
+import mktpromomarc.flotilla.repository.UsuariosRepository;
+import mktpromomarc.flotilla.security.Encoder;
+import mktpromomarc.flotilla.security.Validator;
+import mktpromomarc.flotilla.service.UsuariosService;
 import org.json.JSONObject;
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -15,10 +15,10 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.example.dto.Token;
-import org.example.modelo.Usuarios;
-import org.example.security.JwtGenerator;
-import org.example.security.RecaptchaVerifier;
+import mktpromomarc.flotilla.dto.Token;
+import mktpromomarc.flotilla.modelo.Usuarios;
+import mktpromomarc.flotilla.security.JwtGenerator;
+import mktpromomarc.flotilla.security.RecaptchaVerifier;
 
 @WebServlet(name = "Login", urlPatterns = {"/login"})
 public class LoginController extends HttpServlet {
