+++ src/main/java/mktpromomarc/flotilla/dto/Credentials.java
@@ -1,4 +1,4 @@
-package org.example.dto;
+package mktpromomarc.flotilla.dto;
 
 public class Credentials {
     String token;
